requirements-dev.txt
# Development tools and dependencies
# Install with: pip install -r requirements-dev.txt

# Code formatting and linting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Type checking
mypy>=1.5.0

# Security scanning
bandit>=1.7.5

# Pre-commit hooks
pre-commit>=3.5.0

# Test parallelization
pytest-xdist>=3.5.0

# Async test support (pytest plugin)
pytest-asyncio>=0.23.0

# Development tunneling (expose localhost to internet for webhook testing)
# Used for: Telegram webhook development, API testing with external services
# Production should use proper HTTPS domain (see config/telegram.yaml.example)
pyngrok>=7.0.0,<8.0.0

# Install main requirements too
-r requirements.txt
